Refactor according to PR feedback
